+++ SpaceShooterGame.cpp
@@ -49,9 +49,9 @@ void SpaceShooterGame::onCreate()
 
 	for(UINT i = 0; i < 200; i++)
 	{
-		this->m_asteroids_pos[i] = Vector3(rand() % 4000 + (-2000), rand() % 4000 + (-2000), rand() % 4000 + (-2000));
+		this->m_asteroids_pos[i] = Vector3((float)(rand() % 4000 + (-2000)), (float)(rand() % 4000 + (-2000)), (float)(rand() % 4000 + (-2000)));
 		this->m_asteroids_rot[i] = Vector3((rand() % 628) / 100.0f, (rand() % 628) / 100.0f, (rand() % 628) / 100.0f);
-		float scale = rand() % 20 + (6);
+		float scale = (float)(rand() % 20 + (6));
 		this->m_asteroids_scale[i] = Vector3(scale, scale, scale);
 	}
 
@@ -218,8 +218,8 @@ void SpaceShooterGame::onMouseMove(const POINT &mouse_pos)
 	int width = (win_size.right - win_size.left);
 	int height = (win_size.bottom - win_size.top);
 
-	this->m_delta_mouse_x = (int)(mouse_pos.x - (win_size.left + (width / 2.0f)));
-	this->m_delta_mouse_y = (int)(mouse_pos.y - (win_size.top + (height / 2.0f)));
+	this->m_delta_mouse_x = (float)(mouse_pos.x - (win_size.left + (width / 2.0f)));
+	this->m_delta_mouse_y = (float)(mouse_pos.y - (win_size.top + (height / 2.0f)));
 
 	InputSystem::get()->setCursorPos({ win_size.left + (LONG)(width / 2.0f), win_size.top + (LONG)(height / 2.0f) });
 }
